Refactored WsEvent
